more progress

+++ backend/internal/storage/postgres/messages.go
@@ -3,18 +3,19 @@ package postgres
 import (
 	"context"
 	"go-chat/internal/models"
+	"go-chat/internal/xerrors"
 
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
 )
 
 func (p *Postgres) CreateMessage(ctx context.Context, message models.Message) error {
-	const query string = `INSERT INTO messages (id, room_id, created_at, author, content) VALUES ($1, $2, $3, auth.uid(), $4)`
-	// TODO: remove all usage of auth.uid() - its for RLS. i will be manually handling authorization with the user id i get from the jwt
+	const query string = `INSERT INTO messages (id, room_id, created_at, author, content) VALUES ($1, $2, $3, $4, $5)`
 	if _, err := p.pool.Exec(ctx, query,
 		message.Id,
 		message.RoomId,
 		message.CreatedAt,
+		message.Author,
 		message.Content,
 	); err != nil {
 		return err
@@ -42,10 +43,18 @@ func (p *Postgres) GetMessagesByRoomId(ctx context.Context, roomId uuid.UUID) ([
 	return messages, nil
 }
 
-func (p *Postgres) DeleteMessageById(ctx context.Context, messageId uuid.UUID) error {
-	const query string = `DELETE FROM messages WHERE id = $1`
-	if _, err := p.pool.Exec(ctx, query, messageId); err != nil {
+func (p *Postgres) DeleteMessageById(ctx context.Context, messageId uuid.UUID, userId uuid.UUID) error {
+	const query string = `DELETE FROM messages WHERE id = $1 AND author = $2`
+	ct, err := p.pool.Exec(ctx, query, messageId, userId)
+	if err != nil {
 		return err
+	}
+
+	if ct.RowsAffected() == 0 {
+		return xerrors.NotFoundError("message", map[string]string{
+			"id":     messageId.String(),
+			"author": userId.String(),
+		})
 	}
 
 	return nil
